+++ app/modules/movement_panel.py
@@ -129,7 +129,7 @@ def movement_server(input, output, session, dashboard_data_rv):
         controls=[
             {"type": "navigation", "position": "top-right"},
             {"type": "fullscreen", "position": "top-right"},
-            legend_control(position="bottom-left", show_default=True, show_checkbox=True),
+            legend_control(position="bottom-left", show_default=False, show_checkbox=True),
         ],
         tooltip={
             "html": "<b>{cell_id}</b>",